Improve the way ambient types are referenced.
