+++ build-tests/hashed-folder-copy-plugin-test/src/index.ts
@@ -1,7 +1,3 @@
-import type { IRequireFolderOptions } from '@rushstack/hashed-folder-copy-plugin';
-
-declare function requireFolder(options: IRequireFolderOptions): string;
-
 const ASSETS_BASE_URL: string = requireFolder({
   outputFolder: 'assets_[hash]',
   sources: [
