+++ webpack/hashed-folder-copy-plugin/.gitignore
@@ -1,0 +1,1 @@
+/ambientTypes.d.ts
