+++ webpack/hashed-folder-copy-plugin/config/heft.json
@@ -1,0 +1,25 @@
+{
+  "extends": "@rushstack/heft-node-rig/profiles/default/config/heft.json",
+
+  "eventActions": [
+    {
+      "actionKind": "deleteGlobs",
+      "heftEvent": "clean",
+      "actionId": "defaultClean",
+      "globsToDelete": ["ambientTypes.d.ts"]
+    },
+
+    {
+      "actionKind": "copyFiles",
+      "heftEvent": "bundle",
+      "actionId": "copyAmbientTypesToRoot",
+      "copyOperations": [
+        {
+          "sourceFolder": "src",
+          "destinationFolders": ["."],
+          "includeGlobs": ["ambientTypes.d.ts"]
+        }
+      ]
+    }
+  ]
+}
