+++ webpack/hashed-folder-copy-plugin/src/index.ts
@@ -1,8 +1,4 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-export {
-  IRequireFolderOptions,
-  IRequireFolderSource,
-  HashedFolderCopyPlugin
-} from './HashedFolderCopyPlugin';
+export { HashedFolderCopyPlugin } from './HashedFolderCopyPlugin';
